# Conflicts:
#	src/kantinesimulatie/kantine/Kassa.java

+++ src/kantinesimulatie/klant/KortingskaartHouder.java
@@ -7,7 +7,7 @@ public interface KortingskaartHouder {
     /**
      * Methode om kortingspercentage op te vragen
      */
-    public double geefKortingsPercentage();
+    public BigDecimal geefKortingsPercentage();
 
     /**
      * Methode om op te vragen of er maximum per keer aan de korting zit
